releasing 3.1.3

+++ org.caleydo.view.entourage/src/org/caleydo/view/entourage/GLPathwayGridLayout.java
@@ -215,10 +215,10 @@ public class GLPathwayGridLayout implements IGLLayout {
 		for (PathwayColumn column : copyColumns) {
 			float columnWidth = 0;
 			if (minTotalLevel1Size == 0) {
-				columnWidth = ((float) column.getMinWidth() / (float) totalFixedSize) * getFreeHorizontalSpace(w);
+				columnWidth = (column.getMinWidth() / totalFixedSize) * getFreeHorizontalSpace(w);
 			} else {
 				if (level1Columns.contains(column)) {
-					columnWidth = ((float) column.getMinWidth() / (float) minTotalLevel1Size)
+					columnWidth = (column.getMinWidth() / minTotalLevel1Size)
 							* (getFreeHorizontalSpace(w) - totalFixedSize);
 				} else {
 					columnWidth = column.getMinWidth();
@@ -450,10 +450,10 @@ public class GLPathwayGridLayout implements IGLLayout {
 			return minHeight;
 		}
 
-		public int getMinWidth() {
-			int maxMinWidth = 0;
+		public float getMinWidth() {
+			float maxMinWidth = 0;
 			for (GLPathwayWindow window : windows) {
-				int minWidth = window.getMinWidth();
+				float minWidth = window.getMinWidth();
 				if (minWidth > maxMinWidth)
 					maxMinWidth = minWidth;
 			}
@@ -484,8 +484,8 @@ public class GLPathwayGridLayout implements IGLLayout {
 				if (level4Windows.contains(window)) {
 					windowHeight = window.getMinHeight();
 				} else {
-					int minHeight = window.getMinHeight();
-					float factor = (float) minHeight / (float) totalOtherLevelMinSize;
+					float minHeight = window.getMinHeight();
+					float factor = minHeight / totalOtherLevelMinSize;
 					windowHeight = factor * (freeSpaceVertical - totalLevel4Size);
 				}
 				windowToElement.get(window).setSize(w, windowHeight);
